Add `.barrier()` EIP-method
* Upgrade to Gradle `2.14.1` and some dependencies
* Some `build.gradle` polishing
* Add `BarrierSpec` and `IntegrationFlowDefinition.barrier(...)` fluent API, respectively
* Add `CorrelationHandlerTests` for correlation-specific components
* Move some correlation-specific and XML test to concrete test classes
* Add `PollerSpec.errorChannel(String)` variant
* Fix `IntegrationFlowBeanPostProcessor` to register components from the `SourcePollingChannelAdapterSpec`
